Added new test cases for cursor up command and cursor down in combination with cursor up.  Modified typescript to implement the cursor down and up commands.
Note: this implementation required changing the model so that the
cursor position on the line does not necessarily correspond to any
character on the line.  Only when characters are placed on the line at
that position are spaces added to pad the characters on the line to
fill up to the correct position.